Compiles with Arduino IDE 2.3.1 and AVRBoard package 1.8.6
SDBaseFile.cpp and .h uses a struct definition "fpos_t".
In newer versions fpos_t is already defined in stdio.h and this will conflict with the struct definition in SDBaseFile.

The problem can be resolved be renaming the name of this struct definition.

With this changes, it compiles and can be uploaded to an original printer board.